+++ .gitignore
@@ -3,6 +3,4 @@
 .vscode/c_cpp_properties.json
 .vscode/launch.json
 .vscode/ipch
-mysdcard/
 data/
-data_backup/
