+++ data_tmpl/README.md
@@ -1,0 +1,1 @@
+Base folder for data directory
