+++ platformio.ini
@@ -16,7 +16,7 @@ build_type = debug
 build_flags = -Isrc/lib
 monitor_filters = esp8266_exception_decoder, colorize
 
-board_build.ldscript = eagle.flash.8m7m.ld
+board_build.ldscript = eagle.flash.8m6m.ld
 board_build.filesystem = littlefs
 ; extra_scripts = LittleFsBuilder.py
 
@@ -33,6 +33,8 @@ lib_deps =
     marcmerlin/FastLED NeoMatrix
     marvinroger/ESP8266TrueRandom
     m5ez/ezTime @ ^0.8.3
+    me-no-dev/ESPAsyncTCP @ ^1.2.2
+    me-no-dev/ESP Async WebServer @ ^1.2.3
     ayushsharma82/AsyncElegantOTA @ ^2.2.5
 
 monitor_speed = 74880
