+++ src/webserver.cpp
@@ -1,5 +1,7 @@
 #include <ESP8266WiFi.h>
-#include <ESP8266WebServer.h>
+#include <ESPAsyncTCP.h>
+#include <ESPAsyncWebServer.h>
+#include <AsyncElegantOTA.h>
 #include <uri/UriBraces.h>
 #include <uri/UriRegex.h>
 #include <ESP8266mDNS.h>
@@ -20,8 +22,7 @@ const char *fsName = "LittleFS";
 FS *fileSystem = &LittleFS;
 LittleFSConfig fileSystemConfig = LittleFSConfig();
 
-ESP8266WebServer server(80); // Create a webserver object that listens for HTTP request on port 80
-// WebSocketsServer webSocket(81);    // create a websocket server on port 81
+AsyncWebServer server(80);  // Create a webserver object that listens for HTTP request on port 80
 
 char *mdnsName = "pixelframe"; // Domain name for the mDNS responder
 
@@ -30,44 +31,49 @@ static const char APPLICATION_JSON[] PROGMEM = "application/json";
 static const char FS_INIT_ERROR[] PROGMEM = "FS INIT ERROR";
 static const char FILE_NOT_FOUND[] PROGMEM = "FileNotFound";
 
-////////////////////////////////
-// Utils to return HTTP codes, and determine content-type
+// ////////////////////////////////
+// // Utils to return HTTP codes, and determine content-type
 
-void replyOK()
+void replyOK(AsyncWebServerRequest *request)
 {
-  server.send(200, FPSTR(TEXT_PLAIN), "");
+  AsyncWebServerResponse *response = request->beginResponse(200, FPSTR(TEXT_PLAIN), "");
+  response->addHeader("Access-Control-Allow-Origin", "*");
+  request->send(response);
 }
 
-void replyOKWithMsg(String msg)
+void replyOKWithMsg(AsyncWebServerRequest *request, String msg)
 {
-  server.sendHeader("Access-Control-Allow-Origin", "*");
-  server.send(200, FPSTR(TEXT_PLAIN), msg);
+  AsyncWebServerResponse *response = request->beginResponse(200, FPSTR(TEXT_PLAIN), msg);
+  response->addHeader("Access-Control-Allow-Origin", "*");
+  request->send(response);
 }
 
-void replyOKWithJson(String serializedJson)
+void replyOKWithJson(AsyncWebServerRequest *request, String serializedJson)
 {
-  server.sendHeader("Access-Control-Allow-Origin", "*");
-  server.send(200, FPSTR(APPLICATION_JSON), serializedJson);
+  AsyncWebServerResponse *response = request->beginResponse(200, FPSTR(APPLICATION_JSON), serializedJson);
+  response->addHeader("Access-Control-Allow-Origin", "*");
+  request->send(response);
 }
 
-void replyNotFound(String msg)
+void replyNotFound(AsyncWebServerRequest *request, String msg)
 {
-  server.sendHeader("Access-Control-Allow-Origin", "*");
-  server.send(404, FPSTR(TEXT_PLAIN), msg);
+  AsyncWebServerResponse *response = request->beginResponse(404, FPSTR(TEXT_PLAIN), msg);
+  response->addHeader("Access-Control-Allow-Origin", "*");
+  request->send(response);
 }
 
-void replyBadRequest(String msg)
+void replyBadRequest(AsyncWebServerRequest *request, String msg)
 {
-  Serial.println(msg);
-  server.sendHeader("Access-Control-Allow-Origin", "*");
-  server.send(400, FPSTR(TEXT_PLAIN), msg + "\r\n");
+  AsyncWebServerResponse *response = request->beginResponse(400, FPSTR(TEXT_PLAIN), msg + "\r\n");
+  response->addHeader("Access-Control-Allow-Origin", "*");
+  request->send(response);
 }
 
-void replyServerError(String msg)
+void replyServerError(AsyncWebServerRequest *request, String msg)
 {
-  Serial.println(msg);
-  server.sendHeader("Access-Control-Allow-Origin", "*");
-  server.send(500, FPSTR(TEXT_PLAIN), msg + "\r\n");
+  AsyncWebServerResponse *response = request->beginResponse(500, FPSTR(TEXT_PLAIN), msg + "\r\n");
+  response->addHeader("Access-Control-Allow-Origin", "*");
+  request->send(response);
 }
 
 String getContentType(String filename)
@@ -87,8 +93,9 @@ String getContentType(String filename)
   return "text/plain";
 }
 
-bool handleFileRead(String path)
+bool handleFileRead(AsyncWebServerRequest *request)
 { // send the right file to the client (if it exists)
+  String path = String(request->url());
   Serial.println("[WEBSERVER] handleFileRead: " + path);
   if (path.endsWith("/"))
     path += "index.html";                    // If a folder is requested, send the index file
@@ -98,9 +105,10 @@ bool handleFileRead(String path)
   {                                       // If the file exists, either as a compressed archive, or normal
     if (LittleFS.exists(pathWithGz))      // If there's a compressed version available
       path += ".gz";                      // Use the compressed verion
-    File file = LittleFS.open(path, "r"); // Open the file
-    server.streamFile(file, contentType); // Send it to the client
-    file.close();                         // Close the file again
+
+    // TODO: Not sure how to set content-type
+    request->send(LittleFS, path);
+
     Serial.println(String("\t[WEBSERVER] Sent file: ") + path);
     return true;
   }
@@ -108,7 +116,7 @@ bool handleFileRead(String path)
   return false;
 }
 
-void handleGetFiles(String path)
+void handleGetFiles(AsyncWebServerRequest * request, String path)
 {
   if (path != "/" && !fileSystem->exists(path))
   {
@@ -116,87 +124,95 @@ void handleGetFiles(String path)
   }
 
   Serial.println(String("[WEBSERVER] handleFileList: ") + path);
+
+  // request->addHeader("Access-Control-Allow-Origin", "*");
+
   Dir dir = fileSystem->openDir(path);
   path.clear();
 
-  server.sendHeader("Access-Control-Allow-Origin", "*");
-  // use HTTP/1.1 Chunked response to avoid building a huge temporary string
-  if (!server.chunkedResponseModeStart(200, "application/json"))
-  {
-    server.send(505, F("text/html"), F("HTTP1.1 required"));
-    return;
-  }
+  AsyncWebServerResponse *response = request->beginChunkedResponse("application/json", [&](uint8_t *buffer, size_t maxLen, size_t index) -> size_t {
+    // use the same string for every line
+    String output;
+    output.reserve(64);
+    while (dir.next())
+    {
 
-  // use the same string for every line
-  String output;
-  output.reserve(64);
-  while (dir.next())
-  {
+      if (output.length())
+      {
+        // send string from previous iteration
+        // as an HTTP chunk
+        // sendContent(output);
+        output = ',';
+      }
+      else
+      {
+        output = '[';
+      }
 
-    if (output.length())
-    {
-      // send string from previous iteration
-      // as an HTTP chunk
-      server.sendContent(output);
-      output = ',';
-    }
-    else
-    {
-      output = '[';
-    }
+      output += "{\"type\":\"";
+      if (dir.isDirectory())
+      {
+        output += "dir";
+      }
+      else
+      {
+        output += F("file\",\"size\":\"");
+        output += dir.fileSize();
+      }
 
-    output += "{\"type\":\"";
-    if (dir.isDirectory())
-    {
-      output += "dir";
-    }
-    else
-    {
-      output += F("file\",\"size\":\"");
-      output += dir.fileSize();
-    }
+      output += F("\",\"name\":\"");
+      // Always return names without leading "/"
+      if (dir.fileName()[0] == '/')
+      {
+        output += &(dir.fileName()[1]);
+      }
+      else
+      {
+        output += dir.fileName();
+      }
 
-    output += F("\",\"name\":\"");
-    // Always return names without leading "/"
-    if (dir.fileName()[0] == '/')
-    {
-      output += &(dir.fileName()[1]);
-    }
-    else
-    {
-      output += dir.fileName();
+      output += "\"}";
     }
 
-    output += "\"}";
-  }
+    // send last string
+    output += "]";
+  });
+  request->send(response);
 
-  // send last string
-  output += "]";
-  server.sendContent(output);
-  server.chunkedResponseFinalize();
+
+  // // use HTTP/1.1 Chunked response to avoid building a huge temporary string
+  // if (!server.chunkedResponseModeStart(200, "application/json"))
+  // {
+  //   server.send(505, F("text/html"), F("HTTP1.1 required"));
+  //   return;
+  // }
+
+
+  // server.sendContent(output);
+  // server.chunkedResponseFinalize();
 }
 
-/*
-   Return the list of files in the directory specified by the "dir" query string parameter.
-   Also demonstrates the use of chuncked responses.
-*/
-void handleFileList()
-{
-  if (!server.hasArg("dir"))
-  {
-    return replyBadRequest(F("DIR ARG MISSING"));
-  }
+// /*
+//    Return the list of files in the directory specified by the "dir" query string parameter.
+//    Also demonstrates the use of chuncked responses.
+// */
+// void handleFileList()
+// {
+//   if (!server.hasArg("dir"))
+//   {
+//     return replyBadRequest(F("DIR ARG MISSING"));
+//   }
 
-  String path = server.arg("dir");
+//   String path = server.arg("dir");
 
-  handleGetFiles(path);
-}
+//   handleGetFiles(path);
+// }
 
-void handleNotFound()
+void handleNotFound(AsyncWebServerRequest *request)
 { // if the requested file or page doesn't exist, return a 404 not found error
-  if (!handleFileRead(server.uri()))
+  if (!handleFileRead(request))
   { // check if the file exists in the flash memory (LittleFS), if so, send it
-    server.send(404, "text/plain", "404: File Not Found");
+    request->send(404, "text/plain", "Not found");
   }
 }
 
@@ -218,175 +234,175 @@ void startMDNS()
 void startServer()
 { // Start a HTTP server with a file read handler and an upload handler
 
-  // List directory
-  server.on("/api/files", HTTP_GET, handleFileList);
+//   // List directory
+//   server.on("/api/files", HTTP_GET, handleFileList);
 
-  server.on(UriBraces("/api/show/clock"), HTTP_GET, []() {
-    Serial.println("[WEBSERVER] Receive command - switch to clock");
+//   server.on(UriBraces("/api/show/clock"), HTTP_GET, []() {
+//     Serial.println("[WEBSERVER] Receive command - switch to clock");
 
-    auto ev = new ClockFrameEvent();
-    Orchestrator::Instance()->react(ev);
+//     auto ev = new ClockFrameEvent();
+//     Orchestrator::Instance()->react(ev);
 
-    replyOKWithMsg(F("Switching to clock"));
-  });
+//     replyOKWithMsg(F("Switching to clock"));
+//   });
 
-  server.on(UriBraces("/api/show/off"), HTTP_GET, []() {
-    Serial.println("[WEBSERVER] Receive command - switch to off");
+//   server.on(UriBraces("/api/show/off"), HTTP_GET, []() {
+//     Serial.println("[WEBSERVER] Receive command - switch to off");
 
-    auto ev = new OffEvent();
-    Orchestrator::Instance()->react(ev);
+//     auto ev = new OffEvent();
+//     Orchestrator::Instance()->react(ev);
 
-    replyOKWithMsg(F("Switching to off"));
-  });
+//     replyOKWithMsg(F("Switching to off"));
+//   });
 
-  server.on(UriBraces("/api/show/gif"), HTTP_GET, []() {
-    if (server.hasArg("image"))
-    {
-      String filename = server.arg("image");
-      Serial.print("[WEBSERVER] Receive command - switch to ");
-      Serial.println(filename);
+//   server.on(UriBraces("/api/show/gif"), HTTP_GET, []() {
+//     if (server.hasArg("image"))
+//     {
+//       String filename = server.arg("image");
+//       Serial.print("[WEBSERVER] Receive command - switch to ");
+//       Serial.println(filename);
 
-      uint8_t duration = 10; // TODO: Put default gif duration somewhere more central?
+//       uint8_t duration = 10; // TODO: Put default gif duration somewhere more central?
 
-      if (server.hasArg("duration"))
-      {
-        duration = server.arg("duration").toInt();
-      }
+//       if (server.hasArg("duration"))
+//       {
+//         duration = server.arg("duration").toInt();
+//       }
 
-      auto ev = new SingleGifFrameEvent(duration);
+//       auto ev = new SingleGifFrameEvent(duration);
 
-      ev->filename = std::string(filename.c_str());
-      Orchestrator::Instance()->react(ev);
-    }
-    else
-    {
-      Serial.println("[WEBSERVER] Receive command - switch to random gif");
+//       ev->filename = std::string(filename.c_str());
+//       Orchestrator::Instance()->react(ev);
+//     }
+//     else
+//     {
+//       Serial.println("[WEBSERVER] Receive command - switch to random gif");
 
-      auto ev = new RandomGifFrameEvent();
-      Orchestrator::Instance()->react(ev);
-    }
+//       auto ev = new RandomGifFrameEvent();
+//       Orchestrator::Instance()->react(ev);
+//     }
 
-    replyOKWithMsg(F("Switching to gif"));
-  });
+//     replyOKWithMsg(F("Switching to gif"));
+//   });
 
-  server.on(UriBraces("/api/show/visuals"), HTTP_GET, []() {
-    if (server.hasArg("v"))
-    {
-      String visualArg = server.arg("v");
-      Serial.print("[WEBSERVER] Receive command - switch to visual ");
-      Serial.println(visualArg);
+//   server.on(UriBraces("/api/show/visuals"), HTTP_GET, []() {
+//     if (server.hasArg("v"))
+//     {
+//       String visualArg = server.arg("v");
+//       Serial.print("[WEBSERVER] Receive command - switch to visual ");
+//       Serial.println(visualArg);
 
-      auto ev = new VisualsFrameEvent();
-      auto visual = std::string(visualArg.c_str());
-      ev->visual = visual;
-      Orchestrator::Instance()->react(ev);
+//       auto ev = new VisualsFrameEvent();
+//       auto visual = std::string(visualArg.c_str());
+//       ev->visual = visual;
+//       Orchestrator::Instance()->react(ev);
 
-      replyOKWithMsg(F("Switching to visual frame"));
-    }
-    else
-    {
-      Serial.println("[WEBSERVER] Receive command - switch to visuals frame");
+//       replyOKWithMsg(F("Switching to visual frame"));
+//     }
+//     else
+//     {
+//       Serial.println("[WEBSERVER] Receive command - switch to visuals frame");
 
-      auto ev = new VisualsFrameEvent();
-      ev->visual = "random";
-      Orchestrator::Instance()->react(ev);
+//       auto ev = new VisualsFrameEvent();
+//       ev->visual = "random";
+//       Orchestrator::Instance()->react(ev);
 
-      replyOKWithMsg(F("Switching to random visual frame"));
-    }
-  });
+//       replyOKWithMsg(F("Switching to random visual frame"));
+//     }
+//   });
 
-  server.on(UriBraces("/api/images/{}"), []() {
-    String name = server.pathArg(0);
-    handleFileRead("gifs/" + name);
-  });
+//   server.on(UriBraces("/api/images/{}"), []() {
+//     String name = server.pathArg(0);
+//     handleFileRead("gifs/" + name);
+//   });
 
-  server.on(UriBraces("/api/configuration/basic"), HTTP_GET, []() {
-    Serial.println("[WEBSERVER] GET /configuration/basic");
+//   server.on(UriBraces("/api/configuration/basic"), HTTP_GET, []() {
+//     Serial.println("[WEBSERVER] GET /configuration/basic");
 
-    StaticJsonDocument<200> config;
-    config["brightness"] = matrix_brightness;
-    config["timezone"] = "Europe/Berlin";
+//     StaticJsonDocument<200> config;
+//     config["brightness"] = matrix_brightness;
+//     config["timezone"] = "Europe/Berlin";
 
-    char json_string[200];
-    serializeJson(config, json_string);
+//     char json_string[200];
+//     serializeJson(config, json_string);
 
-    replyOKWithJson(String(json_string));
-  });
+//     replyOKWithJson(String(json_string));
+//   });
 
-  server.on(UriBraces("/api/configuration/basic"), HTTP_PATCH, []() {
-    Serial.println("[WEBSERVER] PATCH /configuration/basic");
+//   server.on(UriBraces("/api/configuration/basic"), HTTP_PATCH, []() {
+//     Serial.println("[WEBSERVER] PATCH /configuration/basic");
 
-    StaticJsonDocument<200> config;
+//     StaticJsonDocument<200> config;
 
-    DeserializationError error = deserializeJson(config, server.arg("plain"));
+//     DeserializationError error = deserializeJson(config, server.arg("plain"));
 
-    if (error)
-    {
-      replyBadRequest(F("Unable to parse body"));
-      return;
-    }
+//     if (error)
+//     {
+//       replyBadRequest(F("Unable to parse body"));
+//       return;
+//     }
 
-    if (config["brightness"] != nullptr) {
-      set_brightness(config["brightness"]);
-    }
+//     if (config["brightness"] != nullptr) {
+//       set_brightness(config["brightness"]);
+//     }
 
-    replyOKWithMsg(F("Updating basic configuration"));
-  });
+//     replyOKWithMsg(F("Updating basic configuration"));
+//   });
 
-  server.on(UriBraces("/api/configuration/wifi"), HTTP_GET, []() {
-    Serial.println("[WEBSERVER] GET api/configuration/wifi");
+//   server.on(UriBraces("/api/configuration/wifi"), HTTP_GET, []() {
+//     Serial.println("[WEBSERVER] GET api/configuration/wifi");
 
-    StaticJsonDocument<200> config;
-    config["ssid"] = wifi_ssid;
+//     StaticJsonDocument<200> config;
+//     config["ssid"] = wifi_ssid;
 
-    char json_string[200];
-    serializeJson(config, json_string);
+//     char json_string[200];
+//     serializeJson(config, json_string);
 
-    replyOKWithJson(String(json_string));
-  });
+//     replyOKWithJson(String(json_string));
+//   });
 
-  server.on(UriBraces("/api/configuration/wifi"), HTTP_PUT, []() {
-    Serial.println("[WEBSERVER] PUT /configuration/wifi");
+//   server.on(UriBraces("/api/configuration/wifi"), HTTP_PUT, []() {
+//     Serial.println("[WEBSERVER] PUT /configuration/wifi");
 
-    StaticJsonDocument<200> config;
+//     StaticJsonDocument<200> config;
 
-    DeserializationError error = deserializeJson(config, server.arg("plain"));
+//     DeserializationError error = deserializeJson(config, server.arg("plain"));
 
-    if (error) {
-      replyBadRequest(F("Unable to parse body"));
-      return;
-    }
+//     if (error) {
+//       replyBadRequest(F("Unable to parse body"));
+//       return;
+//     }
 
-    if (config["ssid"] == nullptr || config["password"] == nullptr) {
-      replyBadRequest("Body must contain SSID and password");
-      return;
-    }
+//     if (config["ssid"] == nullptr || config["password"] == nullptr) {
+//       replyBadRequest("Body must contain SSID and password");
+//       return;
+//     }
 
-    set_wifi(strdup(config["ssid"]), strdup(config["password"]));
+//     set_wifi(strdup(config["ssid"]), strdup(config["password"]));
 
-    replyOKWithMsg(F("Updating wifi configuration"));
-  });
+//     replyOKWithMsg(F("Updating wifi configuration"));
+//   });
 
-  server.on(UriBraces("/api/environment/wifis"), HTTP_GET, []() {
-    Serial.println("[WEBSERVER] GET api/environment/wifis");
+//   server.on(UriBraces("/api/environment/wifis"), HTTP_GET, []() {
+//     Serial.println("[WEBSERVER] GET api/environment/wifis");
 
-    int numberOfNetworks = WiFi.scanNetworks();
+//     int numberOfNetworks = WiFi.scanNetworks();
 
-    StaticJsonDocument<512> config;
+//     StaticJsonDocument<512> config;
 
-    for(int i = 0; i < numberOfNetworks; i++) {
-      config[i]["ssid"] = WiFi.SSID(i);
-      config[i]["signalStrength"] = WiFi.RSSI(i);
-    }
+//     for(int i = 0; i < numberOfNetworks; i++) {
+//       config[i]["ssid"] = WiFi.SSID(i);
+//       config[i]["signalStrength"] = WiFi.RSSI(i);
+//     }
 
-    char json_string[512];
-    serializeJson(config, json_string);
+//     char json_string[512];
+//     serializeJson(config, json_string);
 
-    replyOKWithJson(String(json_string));
-  });
+//     replyOKWithJson(String(json_string));
+//   });
 
-  server.on("/api/images", HTTP_GET, []() {
-    handleGetFiles("/gifs");
+  server.on("/api/images", HTTP_GET, [](AsyncWebServerRequest * request) {
+    handleGetFiles(request, "/gifs");
   });
 
   server.onNotFound(handleNotFound); // if someone requests any other file or page, go to function 'handleNotFound'
@@ -399,11 +415,12 @@ void startServer()
 void setup_webserver()
 {
   startMDNS();   // Start the mDNS responder
+  AsyncElegantOTA.begin(&server);    // Start ElegantOTA
   startServer(); // Start a HTTP server with a file read handler and an upload handler
 }
 
 void webserver_loop()
 {
-  server.handleClient(); // run the server
+  AsyncElegantOTA.loop();
   MDNS.update();
 }
